Parametrize step implementations
Refactored some step implementations to take more parameters.
Moved place_submarine method to uw game daos.
