+++ features/underwater_features/under_init_game.feature
@@ -34,20 +34,20 @@ Feature: Init Under Game
     Scenario: Place a submarine
         Given the user 'player1' is in a game of dimension '6'x'6' with visitor
         And the user 'player1' chose 'Saukko' submarine
-        When they choose the position '2','4' and direction '0' for their submarine
+        When the user 'player1' chooses the position '2','4' and direction '0'
         Then the submarine is successfully placed
 
     Scenario: Place a submarine in an invalid position
         Given the user 'player1' is in a game of dimension '6'x'6' with visitor
         And the user 'player1' chose 'Saukko' submarine
-        When they choose the position '-2','4' and direction '0' for their submarine
+        When the user 'player1' chooses the position '-2','4' and direction '0'
         Then the system should not allow to place the submarine in that position
     
     Scenario: Place a submarine already placed
         Given the user 'player1' is in a game of dimension '6'x'6' with visitor
         And the user 'player1' chose 'Saukko' submarine
-        When they choose the position '2','4' and direction '0' for their submarine
-        And they choose the position '2','5' and direction '0' for their submarine
+        When the user 'player1' chooses the position '2','4' and direction '0'
+        And the user 'player1' chooses the position '2','5' and direction '0'
         Then the system should not allow to place the submarine again
 
         #Preguntar si este escenario estaría bien 
@@ -60,5 +60,5 @@ Feature: Init Under Game
     #         |   |   | X |   |   |
     #         |   |   |   |   |   |
     #         |   |   |   |   |   |
-    #     When they choose the position '2','2' and direction '2' for their submarine
+    #     When the user 'player1' chooses the position '2','2' and direction '2'
     #     Then the system should not allow to place the submarine
