+++ features/underwater_features/under_ongoing_game.feature
@@ -9,14 +9,7 @@ Feature: Ongoing Under Game
         And the user 'player2' chose 'Nautilus' submarine
 
     Scenario: Rotate and move submarine
-        Given the board is in the following state:
-            | 0 | 1 | 2 | 3 | 4 | 5 |
-            |   |   |   |   |   |   |
-            |   | T |   |   |   |   |
-            |   |   | H |   |   |   |
-            |   |   |   |   |   |   |
-            |   |   |   |   |   |   |
-            |   |   |   |   |   |   |
+        Given the user 'player1' chooses the position '{x:d}','{y:d}' and direction '{d:d}'
         When the user 'player1' rotates the submarine with direction '3'
         And the user 'player1' moves the submarine '2' positions
         Then the board is in the following state:
